Update for Single responsibility
