+++ src/buttonInputs/CustomInput/CustomInput.jsx
@@ -6,7 +6,7 @@ const CustomInput = ({value, setValue, placeholder, secureTextEntry, keyboardTyp
         <View style={styles.container}>
             <TextInput 
             value={value}
-            onChange={setValue}
+            onChangeText={setValue}
             placeholder={placeholder} 
             style={styles.input}
             secureTextEntry={secureTextEntry}
