+++ src/components/navigations/_MainTabNavigation/index.js
@@ -1,3 +1,3 @@
-import MainTabNavigation from './MainTabNavigation';
+// import MainTabNavigation from './MainTabNavigation';
 
-export { MainTabNavigation };
+// export { MainTabNavigation };
