--- src/components/organism/_Profile/Profile.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native';
-
-const Profile = () => {
-  return (
-    <Text>Profile Screen</Text>
-  );
-};
-
-export default Profile;
