--- src/components/organism/_Profile/index.js
@@ -1,3 +1,0 @@
-import Profile from './Profile';
-
-export { Profile };
