+++ src/components/organisms/_WifiInput/WifiInput.jsx
@@ -72,7 +72,7 @@ const WifiInput = ({ route, navigation }) => {
     const handleSubmit = () => {
         connectToDevice(deviceId, deviceDetail.wifiName, deviceDetail.wifiPassword);
         setDeviceDetail({ wifiName: '', wifiPassword: '' });
-        navigation.navigate('Home');
+        navigation.navigate('HomeStack');
     };
 
     return (
